YD-51 Moved mail test logic because it is specific
