+++ appservice/src/intTest/groovy/nu/yona/server/CreateUserOnBuddyRequestTest.groovy
@@ -8,6 +8,9 @@ import spock.lang.Specification
 import spock.lang.Unroll
 import groovy.json.*
 
+import javax.mail.*
+import javax.mail.search.*
+
 class CreateUserOnBuddyRequestTest extends Specification {
 
 	def adminServiceBaseURL = System.properties.'yona.adminservice.url'
@@ -96,7 +99,7 @@ class CreateUserOnBuddyRequestTest extends Specification {
 			}""", richardQuinPassword)
 			richardQuinBobBuddyURL = response.responseData._links.self.href
 			
-			def bobDunnInviteEmail = appService.getMessageFromGmail(bobDunnGmail, bobDunnGmailPassword, beforeRequestDateTime)
+			def bobDunnInviteEmail = getMessageFromGmail(bobDunnGmail, bobDunnGmailPassword, beforeRequestDateTime)
 			def bobDunnInviteURLMatch = bobDunnInviteEmail.content =~ /$appServiceBaseURL[\w\-\/=\?&+]+/
 			assert bobDunnInviteURLMatch
 			bobDunnInviteURL = bobDunnInviteURLMatch.group()
@@ -194,5 +197,62 @@ class CreateUserOnBuddyRequestTest extends Specification {
 		then:
 			responseRichard.status == 200
 			responseBob.status == 200
+	}
+	
+	def getMessageFromGmail(login, password, sentAfterDateTime)
+	{
+		def host = "imap.gmail.com";
+	
+		Properties props = new Properties()
+		props.setProperty("mail.store.protocol", "imap")
+		props.setProperty("mail.imap.host", host)
+		props.setProperty("mail.imap.port", "993")
+		props.setProperty("mail.imap.ssl.enable", "true");
+		def session = Session.getDefaultInstance(props, null)
+		def store = session.getStore("imap")
+		
+		def inbox
+		def lastMessage
+		try
+		{
+			println "Connecting to imap store"
+			store.connect(host, login, password)
+			inbox = store.getFolder("INBOX")
+			inbox.open(Folder.READ_WRITE)
+			return getMessageFromInbox(inbox, sentAfterDateTime)
+		}
+		finally
+		{
+			 if(inbox) 
+			 {
+			    inbox.close(true)
+			 }
+			 store.close()
+		}
+	}
+	
+	def getMessageFromInbox(inbox, sentAfterDateTime)
+	{
+		def maxWaitSeconds = 15
+		def pollSeconds = 3
+		def retries = maxWaitSeconds / pollSeconds
+		sleep(100)
+		for (def i = 0; i <retries; i++) 
+		{
+			println "Reading messages from inbox"
+			def messages = inbox.search(
+				//new ReceivedDateTerm(ComparisonTerm.GT,sentAfterDateTime))
+				new FlagTerm(new Flags(Flags.Flag.SEEN), false))
+			if(messages.size() > 0)
+			{
+				def lastMessage = messages[0]
+				def lastMessageMap = [subject:lastMessage.getSubject(), content:lastMessage.getContent()]
+				println "Found message: " + lastMessageMap.subject
+				println lastMessageMap.content
+				return lastMessageMap
+			}
+			sleep(pollSeconds * 1000)
+		}
+		assert false
 	}
 }
